Move Muple to hail.utils
